backend/requirements.txt
asciinema==2.0.2
bech32==1.2.0
BTrees==4.9.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
ecdsa==0.17.0
Flask==2.0.2
future==0.18.2
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.39.0
grpcio-tools==1.39.0
idna==3.2
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
-e git+ssh://git@github.com/kolliderhq/kollider-api-client.git@ae815684e4e6339df1922ecec0adfff27596879b#egg=kollider_api_client
lnurl==0.3.6
MarkupSafe==2.0.1
persistent==4.7.0
protobuf==3.17.3
pycparser==2.21
pydantic==1.9.0
Pygments==2.10.0
pyzmq==22.2.1
requests==2.26.0
rich==10.7.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
